+++ src/MessageEngine.cpp
@@ -58,7 +58,7 @@ string MessageEngine::get(string key) {
 string MessageEngine::get(string key, int i) {
 	string msg = messages[key];
 	if (msg == "") msg = key;
-	int index = msg.find("%d");
+	size_t index = msg.find("%d");
 	if (index != string::npos) msg = msg.replace(index, 2, str(i));
 	return msg;
 }
@@ -66,7 +66,7 @@ string MessageEngine::get(string key, int i) {
 string MessageEngine::get(string key, string s) {
 	string msg = messages[key];
 	if (msg == "") msg = key;
-	int index = msg.find("%s");
+	size_t index = msg.find("%s");
 	if (index != string::npos) msg = msg.replace(index, 2, s);
 	return msg;
 }
@@ -74,7 +74,7 @@ string MessageEngine::get(string key, string s) {
 string MessageEngine::get(string key, int i, string s) {
 	string msg = messages[key];
 	if (msg == "") msg = key;
-	int index = msg.find("%d");
+	size_t index = msg.find("%d");
 	if (index != string::npos) msg = msg.replace(index, 2, str(i));
 	index = msg.find("%s");
 	if (index != string::npos) msg = msg.replace(index, 2, s);
@@ -84,7 +84,7 @@ string MessageEngine::get(string key, int i, string s) {
 string MessageEngine::get(string key, int i, int j) {
 	string msg = messages[key];
 	if (msg == "") msg = key;
-	int index = msg.find("%d");
+	size_t index = msg.find("%d");
 	if (index != string::npos) msg = msg.replace(index, 2, str(i));
 	index = msg.find("%d");
 	if (index != string::npos) msg = msg.replace(index, 2, str(j));
